Implementation du zoomIn elastique
Ajout des appels librairies en C#, ajout des déclarations
et implémentation des librairies permettant le zoomIn.
Le code d'interface graphique est mis à jour afin
de pouvoir supporter le zoomIn avec la souris

+++ Commun/Utilitaire/Vue/ProjectionOrtho.h
@@ -63,7 +63,7 @@ namespace vue {
 		/// Translater la fenêtre virtuelle d'un vecteur
 		void translater(const glm::ivec2& deplacement);
 		/// Centrer la fenêtre virtuelle sur un point
-		void centrerSurPoint(const glm::ivec2& pointCentre);
+		void centrerSurPoint(const glm::dvec2& pointCentre);
 
 		/// Obtenir les coordonnées de la fenêtre virtuelle.
 		inline void obtenirCoordonneesFenetreVirtuelle(
